fix: 修复审计发现的启动失败、安全与解析问题
- 链入口节点已作为主节点或选择器成员构建过时，先复用其全局 tag 再写上一跳的
  detour，否则引用不存在的 g-<id>，sing-box 以 dependency not found 拒绝启动
- 规则库与 ca.pem 从外部存储迁到 filesDir/assets：Android 10 及以下任意持有
  WRITE_EXTERNAL_STORAGE 的应用都能改写外部目录，等于可植入根证书或替换路由库；
  首次启动搬迁旧文件，AssetsActivity 新增 ca.pem 导入并提示重启生效
- Clash API 生成每安装随机 secret，面板 URL 追加 ?secret= 并改用 /ui/（sing-box
  的 /ui 跳转会丢 query）；导出配置与崩溃报告不带 secret
- 分享链接解析改按前缀换 scheme，ss toUri 只替换首个 port/；Quantumult CSV 的
  Host/obfs-path 各自判存在再取值
- SubscriptionBean Kryo v4：expiryDate、lastUpdated 改 Long，lastUpdated 原位保留
  截断 int 供旧版本读，精确值追加在末尾，降级不错位
- snakeyaml 1.30 -> 1.33 修复构造 YAML 的 DoS，code point 上限放宽到 32 MB
- 订阅文件改用 OpenDocument 并持久化授权；clash tuic 的 token（v4）节点导入即丢弃；
  节点搜索按分组顺序过滤；logcat -c 子进程回收

+++ app/src/main/java/io/nekohasekai/sagernet/bg/SubscriptionUpdater.kt
@@ -93,7 +93,7 @@ object SubscriptionUpdater {
                 if (subscriptions.isNotEmpty()) for (profile in subscriptions) {
                     val subscription = profile.subscription!!
 
-                    if (((System.currentTimeMillis() / 1000).toInt() - subscription.lastUpdated) < subscription.autoUpdateDelay * 60) {
+                    if ((System.currentTimeMillis() / 1000 - subscription.lastUpdated) < subscription.autoUpdateDelay * 60) {
                         Logs.d("work: not updating " + profile.displayName())
                         continue
                     }
